+++ crates/canic-core/src/api/lifecycle/root.rs
@@ -1,5 +1,7 @@
 use crate::{
-    config::schema::ConfigModel, dto::fleet_subnet_root::FleetSubnetRootInitArgs, lifecycle,
+    config::{RoleRuntimeAuthority, schema::ConfigModel},
+    dto::fleet_subnet_root::FleetSubnetRootInitArgs,
+    lifecycle,
 };
 
 ///
@@ -12,6 +14,7 @@ impl LifecycleApi {
     pub fn init_root_canister_before_bootstrap(
         args: FleetSubnetRootInitArgs,
         embedded_release_build_id: Option<&str>,
+        runtime_authority: RoleRuntimeAuthority,
         config: ConfigModel,
         config_source: &str,
         config_path: &str,
@@ -19,6 +22,7 @@ impl LifecycleApi {
         lifecycle::init::root::init_root_canister_before_bootstrap(
             args,
             embedded_release_build_id,
+            runtime_authority,
             config,
             config_source,
             config_path,
@@ -27,11 +31,13 @@ impl LifecycleApi {
 
     #[must_use]
     pub fn post_upgrade_root_canister_before_bootstrap(
+        runtime_authority: RoleRuntimeAuthority,
         config: ConfigModel,
         config_source: &str,
         config_path: &str,
     ) -> bool {
         lifecycle::upgrade::root::post_upgrade_root_canister_before_bootstrap(
+            runtime_authority,
             config,
             config_source,
             config_path,